ft: Created WalletDTO classes and skeletal DTO classes for other services.
- Changed balance type from double to BigDecimal in AbstractWallet class.

+++ src/main/java/com/banking/wallet/WalletOperations.java
@@ -1,5 +1,7 @@
 package com.banking.wallet;
 
+import java.math.BigDecimal;
+
 public interface WalletOperations {
 
     double MAX_BALANCE = 50_000.0;
@@ -13,5 +15,5 @@ public interface WalletOperations {
 
     String getWalletId();
 
-    double getBalance();
+    BigDecimal getBalance();
 }
